Add support for Apache Pulsar as Integration
This adds a (publish only) support for Apache Pulsar as an Integration
alternative to MQTT or Kafka.

Apache Pulsar is another "cloud native" distributed messaging and
streaming platform, and this version is kept simple by using pulsar-rs.

Oauth authentication has not been tested, and is only "implemented" as a
pass-through to the library. Deeper integration that involve passing
schemas and more, as well as tracking per-topic producers has been left
as a future improvement.

The current support is to be considered for testing and evaluation
purpouses.

We implement:

- Basic integration test via container
- Configuration options
- Json and Protobuf serialization

Signed-off-by: D.S. Ljungmark <spider@skuggor.se>